+++ client-java/src/main/java/com/kevinguanchedarias/sqs/consumer/TextConsumer.java
@@ -3,6 +3,9 @@
  */
 package com.kevinguanchedarias.sqs.consumer;
 
+import com.kevinguanchedarias.sqs.Message;
+import com.kevinguanchedarias.sqs.TextMessage;
+
 /**
  * 
  * @since 1.0.0
@@ -20,6 +23,16 @@ public class TextConsumer extends AbstractConsumer<String> implements Consumer<S
 	@Override
 	protected String transformResult(String body) {
 		return body.substring(1).substring(0, body.length() - 1);
+	}
+
+	/*
+	 * (non-Javadoc)
+	 * 
+	 * @see com.kevinguanchedarias.sqs.consumer.AbstractConsumer#getMessageClass()
+	 */
+	@Override
+	protected Class<? extends Message<String>> getMessageClass() {
+		return TextMessage.class;
 	}
 
 }